Get up and down ygg

+++ Europium/Dtos/YggTorrentAccount.cs
@@ -3,6 +3,6 @@
 public class YggTorrentAccount
 {
     public decimal Ratio { get; set; }
-    public long Up { get; set; }
-    public long Down { get; set; }
+    public decimal Up { get; set; }
+    public decimal Down { get; set; }
 }
